feat(cli): add option `renderInSubprocess` to avoid polluting global scope
