+++ .changeset/proud-houses-love.md
@@ -1,0 +1,5 @@
+---
+'@pintora/cli': patch
+---
+
+feat(cli): add option `renderInSubprocess` to avoid polluting global scope
