+++ packages/pintora-cli/src/sameprocess-render.ts
@@ -1,0 +1,6 @@
+import type { CLIRenderOptions } from './render-impl'
+import { render as renderImpl } from './render-impl'
+
+export function renderInCurrentProcess(opts: CLIRenderOptions) {
+  return renderImpl(opts)
+}
